3.1.7. Refactor of List package and LinkedListOfIntegeres (remove Node Class)
